Header and english
Updated the header and translated all to english.

+++ Webshoppen2/Webshoppen2/Models/Product.cs
@@ -10,7 +10,7 @@ namespace Webshoppen2.Models
         public int? SupplierId { get; set; }
         public int? CategoryId { get; set; }
         public string? InfoText { get; set; }
-        public float? Price { get; set; }
+        public double? Price { get; set; }
         public int? UnitsInStock { get; set; }
         public bool? ChosenProduct { get; set; }
